Collaborative Game Design Project
The "Tower of Death" was a group project created through a Game Design 1 class in GameMaker where the player is to fight their way up the tower and defeat the final boss. My contributions to the project included importing the sprite sheets provided by the artists, cutting the sprite sheets into frames for animated sprites, and also Player gameplay (movement, attack, health, and respawn) as well the exit code (Press ESC to exit at any time). This project also allowed me to work with GitHub collaboratively.